pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.8.3",
    "dataclasses>=0.8",
    "livekit-agents[assemblyai,cartesia,deepgram,google,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "openai-whisper>=20250625",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
]
